[REF] ui_sheet: extract getters into helpers
Task: 3240138
