+++ src/plugins/ui_feature/ui_sheet.ts
@@ -4,7 +4,12 @@ import {
   NEWLINE,
   PADDING_AUTORESIZE_HORIZONTAL,
 } from "../../constants";
-import { computeIconWidth, computeTextWidth, positions } from "../../helpers/index";
+import {
+  computeIconWidth,
+  computeTextWidth,
+  getMultiLineText,
+  positions,
+} from "../../helpers/index";
 import { localizeFormula } from "../../helpers/locale";
 import { Command, CommandResult, LocalCommand, UID } from "../../types";
 import {
@@ -115,62 +120,7 @@ export class SheetUIPlugin extends UIPlugin {
   getCellMultiLineText(position: CellPosition, width: number | undefined): string[] {
     const style = this.getters.getCellStyle(position);
     const text = this.getters.getCellText(position, this.getters.shouldShowFormulas());
-
-    const brokenText: string[] = [];
-    for (const line of text.split("\n")) {
-      const words = line.split(" ");
-
-      if (!width) {
-        brokenText.push(line);
-        continue;
-      }
-
-      let textLine = "";
-      let availableWidth = width;
-
-      for (let word of words) {
-        const splitWord = this.splitWordToSpecificWidth(this.ctx, word, width, style);
-        const lastPart = splitWord.pop()!;
-        const lastPartWidth = computeTextWidth(this.ctx, lastPart, style);
-
-        // At this step: "splitWord" is an array composed of parts of word whose
-        // length is at most equal to "width".
-        // Last part contains the end of the word.
-        // Note that: When word length is less than width, then lastPart is equal
-        // to word and splitWord is empty
-
-        if (splitWord.length) {
-          if (textLine !== "") {
-            brokenText.push(textLine);
-            textLine = "";
-            availableWidth = width;
-          }
-          splitWord.forEach((wordPart) => {
-            brokenText.push(wordPart);
-          });
-          textLine = lastPart;
-          availableWidth = width - lastPartWidth;
-        } else {
-          // here "lastPart" is equal to "word" and the "word" size is smaller than "width"
-          const _word = textLine === "" ? lastPart : " " + lastPart;
-          const wordWidth = computeTextWidth(this.ctx, _word, style);
-
-          if (wordWidth <= availableWidth) {
-            textLine += _word;
-            availableWidth -= wordWidth;
-          } else {
-            brokenText.push(textLine);
-            textLine = lastPart;
-            availableWidth = width - lastPartWidth;
-          }
-        }
-      }
-
-      if (textLine !== "") {
-        brokenText.push(textLine);
-      }
-    }
-    return brokenText;
+    return getMultiLineText(this.ctx, text, style, width);
   }
 
   /**
@@ -238,32 +188,6 @@ export class SheetUIPlugin extends UIPlugin {
     const cellsPositions = positions(this.getters.getColsZone(sheetId, index, index));
     const sizes = cellsPositions.map((position) => this.getCellWidth({ sheetId, ...position }));
     return Math.max(0, ...sizes);
-  }
-
-  private splitWordToSpecificWidth(
-    ctx: CanvasRenderingContext2D,
-    word: string,
-    width: number,
-    style: Style
-  ): string[] {
-    const wordWidth = computeTextWidth(ctx, word, style);
-    if (wordWidth <= width) {
-      return [word];
-    }
-
-    const splitWord: string[] = [];
-    let wordPart = "";
-    for (let l of word) {
-      const wordPartWidth = computeTextWidth(ctx, wordPart + l, style);
-      if (wordPartWidth > width) {
-        splitWord.push(wordPart);
-        wordPart = l;
-      } else {
-        wordPart += l;
-      }
-    }
-    splitWord.push(wordPart);
-    return splitWord;
   }
 
   /**
